pyproject.toml
[project]
name = "airflow-toolkit"
version = "0.2.0"
description = "Librairie open source d'outils, opérateurs et hooks pour Apache Airflow."
authors = [{ name = "Anis Salah-Bey" }]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
  "apache-airflow>=2.9.0",
  "clickhouse-driver>=0.2.9",
  "requests>=2.31.0",
  "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "black",
  "ruff",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
